WIP refactoring rendering, SONAR requirements

+++ core/mesh/src/mesh3d.cpp
@@ -390,13 +390,15 @@ const std::vector<QColor> &Mesh3d::getColors() const { return colorTableVec; }
 
 QVector3D Mesh3d::getOffset() const {
 
-  return QVector3D(
-      -static_cast<float>(image3_->image()->vx) * image3_->xdim() / 2.0f -
-          image3_->image()->tx,
-      -static_cast<float>(image3_->image()->vy) * image3_->ydim() / 2.0f -
-          image3_->image()->ty,
-      -static_cast<float>(image3_->image()->vz) * image3_->zdim() / 2.0f -
-          image3_->image()->tz);
+  return QVector3D(-static_cast<float>(image3_->image()->vx) *
+                           static_cast<float>(image3_->xdim()) / 2.0f -
+                       image3_->image()->tx,
+                   -static_cast<float>(image3_->image()->vy) *
+                           static_cast<float>(image3_->ydim()) / 2.0f -
+                       image3_->image()->ty,
+                   -static_cast<float>(image3_->image()->vz) *
+                           static_cast<float>(image3_->zdim()) / 2.0f -
+                       image3_->image()->tz);
 }
 
 } // namespace sme::mesh
